requirements.txt
streamlit
paddlepaddle
paddleocr
opencv-python
pillow
matplotlib
pandas
numpy